Add experimental option to make the gevent driver stop switching (some) once it takes locks.
This defaults to off, but can be turned on with env var RELSTORAGE_GEVENT_NO_SWITCH_WHEN_LOCKED or by setting an attribute on the driver class.

This should actually return the level of switching (during the cricital phase) back to what it was before #321.
